Update patch set 1
Patch Set 1:

This seems like a bigger feature... is there a design doc somewhere to show where it's supposed to go?

I don't understand why this is separate from VbExLegacy(). The idea with Ctrl+L was that it boots any unsigned payload in the RW_LEGACY section, it doesn't have to be SeaBIOS, you can just put U-Boot there (in fact we used to have that for T124 back in the day, we just never had enough time to get it shippable... you can probably reuse some of the plumbing left in the chromeos-bootimage.ebuild for that). Or do you have a use case for having SeaBIOS and U-Boot on the same device? (Even then, I think it would be better to just add a parameter to VbExLegacy() and using something like crossystem dev_legacy_index to differentiate between payloads, rather than duplicating all this plumbing for every new unsigned payload type we want to support.)

Patch-set: 1
CC: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>