Update patch set 2
Patch Set 2:

> @Julius: I was going by Hung-Te's original commit here which is also in master branch: https://crrev.com/c/1112877

That patch affects the installer in the system image, so it has to be in the master branch to be effective (only stuff that goes into the firmware image itself (on SPI flash) is built from the firmware branch). However, both of your changes here are pure firmware changes, so they can be in the branch. And your feature here doesn't directly tie into Hung-Te's patch either.

> My understanding is that since Alt OS implementation may be significantly different outside of the eve-campfire branch, we are keeping most (all) commits to live in eve-campfire branch.

Yes, I think that's exactly the point. First of all we need to make trade-offs that we might otherwise not want to make to be able to do it all in RW firmware, and secondly we also just don't 100% know where this is going yet. That's the whole point of this Eve-only test -- to gain the experience to understand things about the design that don't end up working that well in practice and should be solved differently. Since we don't know which parts that will be yet, I think we should keep them all out of master for now. Including this one.

Patch-set: 2