signer: fix accessory_rwsig
accessory signing wasn't referencing the key specified in the
signing instructions. While this change doesn't fix the issue,
it works around by providing a fallback to use the default key
in the key container.

As a workaround, the container passed in is the one containing
the specified key, so if only one key is present it must be the
specified key.

BUG=chromium:863464
TEST=run locally
BRANCH=None

Change-Id: Ieeca5773f35b7bf92beae8a2192ed6e6fd9008e6