image_singing: Add '--nodefault_rw_root' for make_dev_ssd.sh.
When the root is default RW mounted, we found that the system is more
fragile due to suspend and resume tests.

It is found that in early builds, many projects tend to need minor fixes
in test image rootfs, for example kernel modules or files read by
kernel. Currently the only way to update those files is to reflash whole
images, but that's not very practical when the network in manufacturing
line is pretty slow. It would be better if we can change a single file.

As a result, we want to allow setting the default root mount option when
running make_dev_ssd.sh. The new --nodefault_rw_root allows disabling
rootfs verification but still mounting rootfs as RO, which makes better
chances for system to be stable, and changes can still be made by an
explicit 'mount -o rw,remount /'.

BUG=None
TEST=./make_dev_ssd.sh --remove_rootfs_verification --nodefault_rw_root

Change-Id: Ie2675e25b77e638ba6c3be8e2a2a3887a95582fc