Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I find it sorta odd that we're doing both. If you're already adding the recoverysw_is_virtual property, it would be very easy to make it externally visible in crossystem and then just query that in your initramfs code, rather than querying something else and relying on error behavior.

I added a recoverysw_is_virtual property since we already had one for devsw_virtual (that's not externally visible). It's not really necessary.

Though, if my understanding of the logic is right (it might not be), we only have the RECSW_VIRTUAL bit for VbSharedData version >= 2 -- so it's possible that a potential `crossystem recoverysw_is_virtual` call to return an error on older devices. So, detection of a switch in the initramfs would be something like:

# Device has VbSharedData version >= 2.
if crossystem recoverysw_virtual; then
   # Virtual recovery switch is set to 0; if recoverysw_cur returns without error we definitely have a switch.
   if crossystem recoverysw_virtual?0 && crossystem recoverysw_cur; then
       has_switch=1
   fi
else
  # We have no VbSharedData to know anything about recoverysw_virtual, so assume if recoverysw_cur returns without error, we have a switch.
  if crossystem recoverysw_cur; then
    has_switch=1
fi

Which is more complex than just a `crossystem recoverysw_cur` call. My understanding here might be wrong -- please let me know if it is.

Patch-set: 4
Reviewer: Gerrit User 1286065 <1286065@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>