Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review+1
> 
> > Though, if my understanding of the logic is right (it might not be), we only have the RECSW_VIRTUAL bit for VbSharedData version >= 2 -- so it's possible that a potential `crossystem recoverysw_is_virtual` call to return an error on older devices. So, detection of a switch in the initramfs would be something like:
> 
> Randall would know for sure, but I bet that all devices with VbSharedData version 1 don't use virtual recovery anyway (and those devices are probably EOL by now anyway). I have been here for almost 6 years and I don't recall that bit ever not being there. ;)
> 
> I don't care hugely one way or the other, so feel free to push this if you prefer, but I think the other option would just be a little bit cleaner (because it allows your shell script to check for what you actually want to know, rather than infer that from the failure of a different thing).

It would definitely be cleaner if all the devices we cared about had that bit, and the earliest Chromebox was released in 2014, so that definitely fits within the last 6 years :)

In that case it sounds like it would be OK to rely on the presence of that bit -- it would've been more important if we had a significant number of devices out in the field with both a recovery switch and no keyboard *and* no RECSW_VIRTUAL bit. But it doesn't sound like we do, so I'll revert to exposing that bit instead.

Patch-set: 4
Reviewer: Gerrit User 1286065 <1286065@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>