Update patch set 5
Patch Set 5:

> It would definitely be cleaner if all the devices we cared about had that bit, and the earliest Chromebox was released in 2014, so that definitely fits within the last 6 years :)

That's not correct, Stumpy was released in 2012 (although it looks like it has already been EOLed now). But the problem isn't really with devices that have a recovery button without this flag, its with devices that don't have one but also predate the flag.

Upon closer inspection, it looks like I was actually a little too confident with my earlier statement. The *init* flag for virtual recovery has existed as long as virtual recovery itself (which came some time around 2012), but it seems that the *shared data* version of that flag was only added with CL:182241 in 2014. The init flags don't get passed all the way through to crossystem, unfortunately.

So we do have a bit of a problem here and this probably won't quite work right for pre-2014 Chromebooks. I'm not quite sure what the best solution is... maybe we're lucky and none of the pre-2014 Chromebooks actually have that GPIO hooked up, so a combination of this test and your previous approach should work. We didn't have that many devices back then, so you can just count them... for all Arm boards I can tell that they're fine (the only one with the recovery-gpio property is Nyan_Kitty, and that one has a physical button).

But for x86 I don't really know how that is piped through, and it's likely decided in firmware (not in the kernel image like on Arm) so you'd have to dig through a jungle of old firmware branches. Maybe an easier way is to just try it out: go to http://cautotest/afe/#tab_id=hosts, find the name a device of the platform you need in the "Ready" state (say chromeos1-row1-rack7-host3 for falco), run 'ssh chromeos1-row1-rack7-host3.cros crossystem recoverysw_cur' and see if you get the error. (You're usually supposed to "lock" these devices in the interface before you mess with them, but a single crossystem query is probably not going to affect any running test.)

It looks like the only x86 board families that fall in this window and haven't been EOLed yet are slippy (Haswell) and rambi (Broadwell). You can see the list of boards belonging to these families on GoldenEye (e.g. on https://cros-goldeneye.corp.google.com/chromeos/console/listBuild?milestone=70#/details , all the names that are indented under the bold-printed family name). (If there's a name we don't have any devices in the lab for, it's probably some test board we never released so you can ignore it.)

Patch-set: 5
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>