futility: cmd_update: Add interface to crossystem values
Add `crossystem` and value getters into `system_env`, so the updater can read
environment data using crossystem. With this change, we can implement the
trial-boot update logic:
 - Find mainfw_act
 - Select and update to the other slot
 - Set crossystem fw_try_{next,count} values to try in next boot.

Note crossystem is also in vboot_reference repository (but not in
futility yet) so in future we may be able to invoke it directly without
doing a real fork.

BUG=chromium:875551
TEST=make futil; ./build/futility/futility -i /build/eve/firmware/image.bin
 futility -i IMAGE -t; futility -i IMAGE --wp=0
BRANCH=None

Change-Id: Ie745726107bff416549ba095a3defdd4cc98d32d
Signed-off-by: Hung-Te Lin <hungte@chromium.org>