futility: cmd_update: Allow getting and setting crossystem values
The `crossystem` value internally calls VbGetSystemProperty{String,Int}
and VbGetSystemProperty*. We can call them directly to prevent invoking
crossystem as shell command.

The crossystem values are considered as "system environment values" that
will change updater logic. To allow easier customization and testing in
future, we want to allow overriding all the values in a single place.

`system_env.properties` is the place to hold cached values, and can be
provisioned to override values. The `get_env_property`
functions will try to look at `system_env.properties[i].initialized`;
and call .getter if not set, or return the cached .value directly.

With this change, we can implement the trial-boot update logic:
 - Find mainfw_act
 - Select and update to the other slot
 - Set crossystem fw_try_{next,count} values to try in next boot.

BUG=chromium:875551
TEST=make futil; ./build/futility/futility -i /build/eve/firmware/image.bin
 futility -i IMAGE -t; futility -i IMAGE --wp=0
BRANCH=None

Change-Id: Ie745726107bff416549ba095a3defdd4cc98d32d
Signed-off-by: Hung-Te Lin <hungte@chromium.org>