futility: cmd_update: Check contents before starting to update
In try-boot updating flow, we have to first check if RO content needs to be
changed or not, and do full (ro+ro) update if WP is disabled. Also, before
starting to update RW-A or RW-B, we should also check if the active system
already has same firmware contents.

BUG=chromium:875551
TEST=make futil; ./build/futility/futility -i /build/eve/firmware/image.bin -t
BRANCH=None

Change-Id: Ie2f75f9aab4696c75aedafbf45e418ee98a2a4b4