futility: cmd_update: Add vboot1 updater logic
There are still many devices running vboot1 and we need to support them as
well. There is no way to determine if a firmware is vboot2 or not, so we
can only rely on the system property "fw_vboot2".

If fw_vboot2 is 0, then we should always update section B and compare content
with section A.

BUG=chromium:875551
TEST=make futil; /futility -i /build/eve/firmware/image.bin -t
BRANCH=None

Change-Id: Iefdcb81099914c2183c627a33eb73678d1269bc1
Signed-off-by: Hung-Te Lin <hungte@chromium.org>