futility: cmd_update: Add quirk 'push_legacy_update'
For devices that do not have update tag provisioned in legacy CBFS, we
need a way to push and enforce the updater to complete first migration.

The --quirks push_legacy_update allows overwriting RW_LEGACY even if the
section on existing system was shipped with an old image that does not
have the AU tag.

BUG=chromium:875551
TEST=make futil; tests/futility/run_test_scripts.sh $(pwd)/build/futility
BRANCH=None

Change-Id: I87db067ad134e82bbbdc937bd2880c6731ec892b
Signed-off-by: Hung-Te Lin <hungte@chromium.org>