Update patch set 3
Patch Set 3:

"accessory" is too generic a name.  the fact that the current set of devices use a single key pair with a specific set of key algos/settings is not a requirement that all future parts do the same.  if anything, the current key pair system we're using for these devices lacks features we'd really want to see (like key rotation).

the reason for each of these fragments isn't because we need the key pair to have a specific name (which was a bad idea that you fixed), but because we want a script tied to that specific crypto usage in a specific device.

Patch-set: 3
Reviewer: Gerrit User 1000768 <1000768@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>