Update patch set 4
Patch Set 4: Code-Review+1

The original script's' is used to do `renaming` in the script so the user (infra side), who calls it to generate key-pair, don't be necessary to care about naming.

This CL here is just in another path like we keep the script as the single one with a unique naming - hammer-like then user need to know the naming in order to re-name it. Both of them are fine to me since the difference is we put effort of re-naming to which ones.

Patch-set: 4
Reviewer: Gerrit User 1176029 <1176029@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1