Update patch set 5
Patch Set 5:

I don't think this is a good idea. The CBFS format is internal to coreboot, it has changed before and it may change again. It is not meant to be reimplemented externally.

The tool you're writing here runs in userspace, so I see no reason why you can't just shell out to cbfstool. You're not copy&pasting flashrom code in here either, after all. (This is very different from when I suggested to not shell out to crossystem, because that code is already in the same repository anyway so it is guaranteed to stay up to date.)

Patch-set: 5
CC: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>