Update patch set 2
Patch Set 2:

I think the concept is useful.  But WP is unfortunately a bit more complicated.

For example, if you deassert HW WP, then you can write to SPI flash without a reboot; flashrom will automatically clear SW WP.  But you can't write to STM32 EC without a reboot, because SW WP is set by the option bytes and only takes effect on the next boot.

I also wonder whether we should be a bit more verbose about how Cr50 is controlling WP, for newer systems.  Is it disabling WP because of CCD open/config, or because battery isn't present?

If we want to combine enabling WP, how do we determine the expected range?  Most of the time now it's half the chip, but in the future it may not be.  And for EC with rollback protection (like hammer), sometimes EC-RW is writable and sometimes the rollback info is writable.

Patch-set: 2
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>