Update patch set 5
Patch Set 5:

> I haven't seen the format changed since ChromeOS starts using it.
> Do you mean the extended attributes?

Yeah, there were some more things that changed together with that (e.g. getting rid of master header in some cases, adding FMAP integration). 

> 
> > It is not meant to be reimplemented externally.
> 
> mosys already has a copy. I can use exactly that copy if you don't like re-implementation - although I feel the impl here is much better and cleaner.

Yikes, I didn't know. I don't think that's a good idea either. I care about having stuff duplicated all over the place, not about the implementation itself, so copying that one wouldn't be better in my opinion.

> I do want to reduce external dependency. For example, in the future if someone wants to run the updater and do simulation on a debian (without cbfstool), with this change he only needs to get a working futility, no need of cbfstool. But if we don't have our copy then we have to port cbfstool as well, which is more painful.

But cbfstool is already designed to work on different platforms (I think it even works under Windows), why would that be a big problem?

>  2. I do want to leverage cbfs more in future for flags, and probably even use that as storage of rootkeys and vblocks (to replace the shar chromeos-firmwareupdate today). And to do that, it's better if futility can read cbfs directly.

So you're going to reuse CBFS as a sort of archive format just to pack the firmware image? That sounds pretty weird too, I think, it's not really designed for that. Why not just use zip or tgz or something? It would probably be a good idea if you could sketch out those plans in a small design doc, so it's clearer where you're trying to go.

> Will it help if I revise cbfstool to make it "libcbfstool", and then just link it?

Yeah, I think that would be the clean option if you really don't want to have an external dependency. Then we could use that for mosys and whatever else wants to use it too. That means building futility depends on the coreboot repo, of course, and building coreboot depends on futility... so a fun little dependency mess to figure out at build time, but not impossible.

Patch-set: 5