Update patch set 1
Patch Set 1: Code-Review+2

this looks fine, but to be sure, you found this by using cs/ and not by some build ?  ebuilds should be explicitly setting CC which should be overriding this setting so it'd never be used in our builds.

if our builds are relying on this being set correctly in this Makefile, that's a bug we need to fix in the ebuild itself.

Patch-set: 1
Reviewer: Gerrit User 1000768 <1000768@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2