futility: update: Move updater logic to 'updater.c'
The updater is getting more complicated and we may want to split into
few modules, for example "updater", "quirks", and "host".

The first step is to change cmd_update.c to updater.c (to preserve most
GIT history), and only declaring the DECLARE_FUTIL_COMMAND in
cmd_update.

BUG=chromium:875551
TEST=make futil; tests/futility/run_test_scripts.sh $(pwd)/build/futility
BRANCH=None

Change-Id: Icae37db8720162130cf38767fec14a970cc9899d
Signed-off-by: Hung-Te Lin <hungte@chromium.org>