futility: updater: Add quirk 'daisy_snow_dual_model' for daisy_snow
The target AUE for daisy_snow is 74 or even longer, so we need to get a
better solution to get rid of script based updater customization (and the
painful EXTRA list in updater configuration).

The new quirk 'daisy_snow_dual_model' is assuming the input firmware image
has both daisy_snow x8 and x16 firmware packed into a single image
(because in vboot1, RW_A is identical to RW_B), and will modify A/B
contents according to target system.

BRANCH=None
BUG=chromium:881034
TEST=make futil; tests/futility/run_test_scripts.sh $(pwd)/build/futility
     # Provide a fake mosys and output both MP / MPx16 to:
     futility update -i bios-snow-2695.132.117-rw.bin \
      --quirks daisy_snow_dual_model --emu emu.bin --sys_props 0,0x0000,0

Change-Id: I8af1b6c3117a703aed4da59902aaecb1009101f2
Signed-off-by: Hung-Te Lin <hungte@chromium.org>