vboot: create NVRAM flag to pause before fetching Alt OS hotkey
Previously, it is impossible to programmatically enable/disable
Alt OS mode.  This is because only EC-RW supports the kbatboot
keyboard matrix functionality.  But as part of the campfire
boot flow, the keyboard matrix is retrieved immediately after
jumping into EC-RW.  We need to insert a small pause in order
to allow for running this command, thus simulating the Alt OS
keyboard press hotkey.

BUG=b:117140648
TEST=Manually use crossystem to set alt_os_hotkey_delay_request=1
     Reboot, and wait for the delay to begin
     Run `kbatboot 1 4 1` in EC console
     Check that AP console contains:
       "VbCheckAltOS: Alt OS: hotkey_after_sync=0"

Change-Id: I1305357199d87b80b4edc4e311015106ab07de65