vboot: create NVRAM flag to pause after EC software sync
Previously, it is impossible to programmatically enable/disable
Alt OS mode in eve.  This is because only EC-RW supports the
kbatboot keyboard matrix functionality.  But, as part of the
campfire boot flow, the keyboard matrix is retrieved *immediately*
after jumping into EC-RW.  We need to insert a small pause in
order to allow for some entity (autotest/servo) to send a kbatboot
command, simulating the Alt OS keyboard press hotkey.

BUG=b:117140648
TEST=Manually use crossystem to set post_ec_sync_delay=1
     Reboot, and wait for the delay to begin
     Run `kbatboot 1 4 1` in EC console
     Check that AP console contains:
       "vb2_post_ec_sync_hooks: post_ec_sync_delay 5000 ms..."

Change-Id: I1305357199d87b80b4edc4e311015106ab07de65