crossystem: add alt_os_enabled to show Alt OS state
Some user-space applications need to know whether Alt OS is
currently enabled or disabled.  Add alt_os_enabled to
crossystem as a read-only flag for this purpose.

It is currently based off of reading VBSD_ALT_OS_SHOW_PICKER
from VbSharedDataHeader.  We may want to change that to a
field dedicated to showing Alt OS state in the future
(see b/117195332).

BUG=b:117195332,b:117142023
TEST=emerge-eve vboot_reference && \
     cros deploy --force --board=eve dut vboot_reference

Change-Id: Ic9a120e7d24021eb984d501f09ce4d7b6f85d730