crossystem: add alt_os_cur to show enabled/disabled
Some user-space applications need to know whether Alt OS is
currently enabled or disabled.  Add alt_os_cur to crossystem
as a read-only flag for this purpose.

It is currently based off of reading VBSD_ALT_OS_SHOW_PICKER
from VbSharedDataHeader.  We may want to change that to a
field dedicated to showing Alt OS state in the future
(see b/117195332).

BUG=b:117195332,b:117142023
TEST=emerge-eve vboot_reference && \
     cros deploy --force --board=eve dut vboot_reference

Change-Id: I27fcebda0606552ad9849456846194458686925e