Add a beep type for things that are against policy
There is one more type of beep - a single high beep. This seems to be used
when something is against policy, for want of a better term. Add this as a
new beep type and update all callers.

All beeping now goes through vb2_error_beep(), except for one beep in
vboot_audio.c.

We could move vb2_error_beep() to vboot_audio.c, but the beeps seem to be
a part of the UI rather than the audio system. Of course,
vb2_audio_looping() arguable is also...

BUG=chromium:837018
BRANCH=none
TEST=FEATURES=test emerge-grunt --nodeps vboot_reference

Change-Id: I55807b4548987a621e8bbced97e7710d6cd6d5fb
Signed-off-by: Simon Glass <sjg@chromium.org>