tpmc: Make 'tpmc def' replace the existing space by default
In chromium:895549, we want to have consistent behavior of
'tpmc def' between TPM 1.2 and TPM 2.0.
In TPM 1.2, define space command will undefine the existing space,
and create a new one. So we make the 'tpmc def' act as this by
default.

Also, provide a option for whom may want to define a new space
only if it is not defined yet. It will return TPM error code
at that case.

BUG=chromium:895549
BRANCH=None
TEST=unit test; manually test:
 # AUTHREAD|AUTHWRITE
 tpmc def 0x1020 0x1 0x40004

 tpmc def 0x1020 0x1 0x40004
 # Expected: Success

 tpmc def 0x1020 0x1 0x40004 --no-overwrite
 # Expected: output TPM error for the space is already defined:
 #  The space is existing but --no-overwrite is set.

Change-Id: I9b4e742f2935578443ebcc69e91d0aebc84deed8