Update patch set 1
Patch Set 1: Code-Review+2

(1 comment)

It's a bit odd because it doesn't actually check whether it's allowed... you tell it whether it's allowed, and if you told it it wasn't, it prints an error. But not a big deal.

Patch-set: 1
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2