ec_sync: allow AUX_FW update even when EC is current
this relaxes when we attempt to do auxiliary FW (i.e. TCPC) firmware
updates.  originally, we would only do an update if we also needed to
do an EC firmware update.  now, we'll do the TCPC firmware update if
the firmware version calls for it regarless of whether we also have an
EC firmware update pending.

BUG=b:118690201
BRANCH=firmware-eve-9584.B
TEST=verified certain bit errors in anx3429 firwmare can be corrected

Change-Id: Ib2860d0abca3839632c92106578a863ede9ad414
Signed-off-by: Caveh Jalali <caveh@chromium.org>